tests/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
paho-mqtt==2.1.0
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
sniffio==1.3.1
starlette==0.44.0
tomli==2.2.1
typing_extensions==4.13.2
uvicorn==0.33.0
